+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file contains "important" commits, small ones will probably not be added.
 
+### 200602 - 1 [0.7.7]
+- Redesigned plugins page
+- Fixed, toolbar button weren't focusing
+- Better directories explorer resizer
 
 ### 200601 - 1 [0.7.7]
 - Now you can expand the files explorer from left to right by clicking on top right of it and moving the cursor